// web-previews/src/utils/contentLink/types.ts
export const SYMBOL_FOR_PRIMARY_ENVIRONMENT = '__PRIMARY__';

export type EditUrlInfo = {
  environment: string;
  itemTypeId: string;
  itemId: string;
  fieldPath: string;
};

export type ContentLinkState = {
  clickToEditEnabled: boolean;
  path: string;
  itemIdsPerEnvironment: Record<string, string[]>;
};

export type OnInitData = {
  editUrlRegExp: { source: string; flags: string };
};

// Methods that we expose
export interface WebPreviewsMethods {
  onInit: () => OnInitData;
  onPing: () => void;
  onStateChange: (payload: ContentLinkState) => void;
  openItem: (payload: EditUrlInfo) => void | Promise<void>;
}

// Methods that @datocms/content-link exposes
export interface ContentLinkMethods {
  navigateTo: (payload: { path: string }) => void;
  flashAll: (payload: { scrollToNearestTarget: boolean }) => void;
  flashItem: (payload: {
    itemId: string;
    scrollToNearestTarget: boolean;
  }) => void;
  setClickToEditEnabled: (
    payload:
      | { enabled: true; flash: { scrollToNearestTarget: boolean } }
      | { enabled: false },
  ) => void;
}